% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.amazon.reviews.R
\name{extract.amazon.reviews}
\alias{extract.amazon.reviews}
\title{pulls product reviews from Amazon.com.}
\usage{
extract.amazon.reviews(url, n, filename, sheetname)
}
\arguments{
\item{url}{url of the products second page}

\item{n}{total number of pages.}

\item{filename}{name of the excel file.}

\item{sheetname}{name of the sheet of excel file.}
}
\value{
Dataframe
}
\description{
pull reviews/opinions from Amazon and store in excel sheet.
}
\examples{
extract.amazon.reviews("https://www.amazon.com/Rich-Dad-Poor-Teach-Middle/product-reviews/1612680194/ref=cm_cr_arp_d_paging_btm_2?ie=UTF8&reviewerType=all_reviews&pageNumber=2",3,"reviews","rich_dad_poor_dad")

}
