% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-reviews.R
\name{get_review_col_types}
\alias{get_review_col_types}
\title{Returns a \code{col_spec} for Amazon Customer Reviews files.}
\usage{
get_review_col_types()
}
\value{
\code{col_spec} for an Amazon Customer Reviews dataset.
}
\description{
Returns a \code{col_spec} for Amazon Customer Reviews data
to override the defaults detected by \code{readr::read_tsv}.
}
