\name{parse_float}
\alias{parse_float}
\title{Parse an XMLNode into a float by stripping everything non-numeric
around the first numeric.}
\usage{
parse_float(value)
}
\arguments{
  \item{value}{XMLNode or character.}
}
\value{
numeric that represents the parsed value.
}
\description{
Parse an XMLNode into a float by stripping everything
non-numeric around the first numeric.
}

