% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-reviews.R
\name{read_review_file}
\alias{read_review_file}
\title{Reads an Amazon Customer Reviews file}
\usage{
read_review_file(path)
}
\arguments{
\item{path}{character vector: fully qualified path to the file to be read.}
}
\value{
\code{tbl} with 14 columns:
\itemize{
\item \code{marketplace} (character vector): 2-letter country code of the marketplace where the review was written.
\item \code{customer_id} (character vector): random identifier that can be used to aggregate reviews written by a single author.
\item \code{review_id} (character vector): unique ID of the review.
\item \code{product_id} (character vector): unique Product ID the review pertains to.
\item \code{product_parent} (character vector): random identifier that can be used to aggregate reviews for the same product.
\item \code{product_title} (character vector): title of the product.
\item \code{product_category} (character vector): broad product category that can be used to group reviews (also used to group the dataset into coherent parts).
\item \code{star_rating} (integer): 1-5 star rating of the review.
\item \code{helpful_votes} (integer): number of helpful votes.
\item \code{total_votes} (integer): number of total votes the review received.
\item \code{vine} (logical): TRUE if review was written as part of the Vine program; FALSE otherwise.
\item \code{verified_purchase} (logical): TRUE if review is on a verified purchase; FALSE otherwise.
\item \code{review_headline} (character vector): title of the review.
\item \code{review_body} (character vector): full text content of the review.
\item \code{review_date} (Date): date the review was written.
}
}
\description{
Reads a \href{https://s3.amazonaws.com/amazon-reviews-pds/readme.html}{Amazon Customer Reviews} \href{https://s3.amazonaws.com/amazon-reviews-pds/tsv/index.txt}{dataset} of reviews of \href{https://www.amazon.com/}{Amazon} products, written by Amazon customers.
}
