% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_logical.R
\name{yes_no_to_logical}
\alias{yes_no_to_logical}
\title{Converts strings to logical values}
\usage{
yes_no_to_logical(x)
}
\arguments{
\item{x}{(character vector) strings like "y", "Y", "yes", "YES", "n", "N", "no", or "NO" that represent a logical value.}
}
\value{
vector where non-case-sensitive strings "y", "yes" have been converted to \code{TRUE} and all other values, including \code{NA} or empty string (""), have been converted to \code{FALSE}.
}
\description{
Converts non-case-sensitive strings like "n" or "y" to logical values FALSE or TRUE.
}
